requirements.txt
fastapi~=0.124.2
uvicorn[standard]~=0.38.0

h11~=0.16.0
typing_extensions~=4.15.0
pip~=25.3
attrs~=25.4.0
pydantic~=2.12.5
pydantic_core~=2.41.5
starlette~=0.50.0
websockets~=15.0.1
anyio~=4.12.0
uvloop~=0.22.1
click~=8.3.1
httptools~=0.7.1
watchfiles~=1.1.1
annotated-types~=0.7.0
PyYAML~=6.0.3
python-multipart~=0.0.20
opencv-python~=4.12.0.88
numpy~=2.2.6
pytesseract~=0.3.13
pillow~=12.0.0
pyjwt~=2.7.0
aiohttp~=3.13.2
python-dotenv~=1.2.1